Move routines out of main.py
This makes fuzz introspector resemble more of a library and should make
testing a bit easier.
